Lay down basic algorithm
It's first going to parse the SVG document to a list of commands. This intermediary stage is necessary because that parsing is recursive: It can parse subtrees of the SVG document and then transform them with CSS or SVG transform elements, so it needs to change lists of commands. Perhaps even repeat them if we want to implement symbols.
Next it's going to lay out these commands as g-code, just a simple translation step.
Next it needs to create a scene node containing both layer view data (in order to preview it in Cura) and g-code data (in order to print the thing).