Transfer line width top-down
Instead of generating the commands at the bottom with an incorrect line width, then transforming them as you go up towards the root, we'll now pass this property down towards the children first, then parse with the correct line width immediately.
This is going to be necessary to implement the transform property, because we still want to have the correct resolution after transformation (to prevent generating lines smaller than the maximum resolution).