Implement relative quadratic curves
The same as the absolute one, but relative.
I'm unsure of the specification here what happens when there are multiple sets of parameters. Does the second set become relative to the ending coordinate of the first set, or does it all stay relative to the original start of the sequence of quadratic curves? The specification is unclear on this.