Add ttfquery module
This is a library that can read system fonts from the user. It seems to be a pure Python solution, so hopefully it's going to be cross-platform.
The library is BSD-licensed, so we can publish it here without attribution.
The library depends on fontTools. This is a huge library which we'll also need to include with our distribution.