Add framework for parsing text
A lot of work to be done still. But this yields results, so I'm happy. It suddenly looks a lot more plausible that we can implement this cross-platform. I just hope that we can find a similar solution for OTF fonts or skip the step of ttfquery some day. fontTools supports so many more formats but we're currently limited to TTF fonts.