Fix for loading too many fonts
FreeType has a limited number of slots to store fonts in. That means that we can't just load all fonts on the system and keep them in memory (via the system_fonts dict). We must release them. Therefore we only store the filenames of the files that we found to have the various font family names.