Attempt at correcting the font size
I arrived at the DPI of 362 experimentally, by drawing a rectangle that fit exactly around the box in Firefox, and then adjusting the DPI until the text exactly fits the rectangle. However there seems to be some rounding going on there too because a DPI of 363 seems to be equal to 365 and 362 seems to be equal to 360. However it should be somewhere in between.