Implement text-decoration super-property
I believe that in this case the priority is wrong. It should take text-decoration-style over the style sub-property in the text-decoration property, if both are present. I'll accept that as a known shortcoming though. It's not all that common anyway. Uncommon enough to not be mentioned anywhere in MDN, even.