Fix viewport offset if not 0,0
The width and height of the viewport were correctly aligned so that it would cover the whole build plate, but the X and Y were unused. They are used now.
I can't do this via transformations since the field is parsed only when the SVG tag is parsed, but inheritance is applied before that. In a pre-parsing step I could do something like converting it to a transform tag and such but this is a better solution I think.