Store CSS attributes in named tuple
This is a good way to enforce that all data is complete (not having a missing validation or something somewhere), which was a problem in the past. It also keeps all of the attribute-specific data together so that it's easy to change. There are no (or fewer) special cases then.