Fix row-column order of transformations
Apparently not well tested! When exporting a file from Inkscape I discovered that the transformation of the document in Cura was completely off. This was due to an affine transformation of like 0.4 units which shouldn't affect the document much. But since this was interpreted as a shear it completely transformed the shape.
With this change the transformation seems to be correct.