Fix crash when clicking two buttons
If two close or cancel events are being triggered in quick succession, Cura would crash. It would release the lock twice, which raises an error. Therefore we'll first re-acquire the lock first, but only if necessary (not blocking). This makes sure we always have the lock before releasing it.
It's not 100% thread safe still, but these events are not processed in parallel so we're okay there.