Fix bug where it recognised scientific notation as a path command
There's a number of path commands to describe curves: M, C, L, H, etc. However, E is not one of them. E is used in the scientific notation, e.g. 2.1E-6 denoting 0.0000021. The E was recognised as a path command since it was a letter, thereby breaking off the actual command.

Fixes #14.